backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# Pydantic and validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Document processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
python-markdown==3.5.1
beautifulsoup4==4.12.2
pypandoc==1.12

# Vector databases
pinecone-client==2.2.4
weaviate-client==3.25.3
chromadb==0.4.18

# Search and embeddings
langchain==0.0.335
langchain-openai==0.0.2
openai==1.3.7
sentence-transformers==2.2.2
elasticsearch==8.11.0
faiss-cpu==1.7.4

# Text processing
nltk==3.8.1
spacy==3.7.2
tiktoken==0.5.2

# Async and background tasks
celery==5.3.4
aiofiles==23.2.1
httpx==0.25.2

# Monitoring and observability
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
structlog==23.2.0

# Rate limiting and caching
slowapi==0.1.9
aiocache==0.12.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
typer==0.9.0 